Address review: move reusable components to SharedProfileComponents
Extracts SharedProfileSubscribeButton, EpisodePlayButton, PlayTriangle,
and ScaleButtonStyle to a dedicated file.
